Consider "image" meta tag when parsing meta data
The extract method `get_article_metadata` in `Readability` parses a list
of meta fields. The first part of the method analyzes the list of
accepted `<meta>` tags, for example OpenGraph, Twitter card etc.

Unfortunately this part skips a number of valid tags, when filtering the list of valid tags to read, in particular it filters according to the [`META_PROPERTY_KEYS`](https://github.com/justahero/dom_smoothie/blob/main/src/glob.rs#L172) list. The subsequent checks to find metadata fields in this set will not work. Among the fields that are skipped for example are the `image` tags, e.g. `og:image`.

This change adds the `image` entry to allow the `get_article_metadata`
to find the image URL in the `head` block of the HTML.